Estates command "/estates buy" bugfix
Command now supports float currency so players can no longer buy for
free (by buying 2 blocks at a time)

+++ me/MnC/MnC_SERVER_MOD/Handlers/CommandsHandler.java
@@ -2710,7 +2710,7 @@ public abstract class CommandsHandler
 					sender.sendMessage("Pocet blocku musi byt cislo mensi nez 2147483648.");
 					return;
 				}
-				int price = (int)(numberOfBlocks * 0.2);
+				float price = numberOfBlocks * 0.2f;
 				if(plugin.currencyManager.getBalance(sender.getName()) >= price)
 				{
 					EstateHandler.addAvailableResidenceBlocks(sender.getName(), numberOfBlocks);
